+++ src/js/components/Home/Home.scss
@@ -18,3 +18,7 @@
 .country-div{
     background-color: #ccc;
 }
+.speedy{
+    height: 25%;
+    width: 25%;
+}
